[ADD][1920097]  Analytic Account on SO lines
What is the business use case for this development?
Currently Analytic Account is on SO level and when you create an invoice, it will apply the same account to all the lines.
I need to be able to set the analytic account per sales order line and when you create the invoice it will take what was at the line level.

Specification :
Add analytic account field to the SO Lines. When you create an invoice for an invoiceable line from the SO this analytic account should carry over.